refactor css: form-tabs, text-field, vertical-nav-bar
