+++ src/App.scss
@@ -4,4 +4,6 @@
 @import './stylesheets/cards';
 @import './stylesheets/forms';
 
-@import './components/nav-bar/nav-bar';
+@import './components/vertical-nav-bar/vertical-nav-bar';
+@import './components/forms/text-field/text-field';
+@import './components/forms/form-tabs/form-tabs';
